+++ src/main/java/com/ajousw/spring/domain/navigation/api/NavigationApiFactory.java
@@ -1,7 +1,6 @@
 package com.ajousw.spring.domain.navigation.api;
 
 
-import com.ajousw.spring.domain.exception.ApiNotSupportedException;
 import com.ajousw.spring.domain.navigation.api.info.SafeNumberParser;
 import com.ajousw.spring.domain.navigation.api.info.route.NaverNavigationApiResponse;
 import com.ajousw.spring.domain.navigation.api.info.route.NavigationApiResponse;
@@ -11,6 +10,7 @@ import com.ajousw.spring.domain.navigation.api.provider.NavigationApi;
 import com.ajousw.spring.domain.navigation.api.provider.Provider;
 import com.ajousw.spring.domain.navigation.api.provider.impl.NaverNavigationApi;
 import com.ajousw.spring.domain.navigation.api.provider.impl.OsrmNavigationApi;
+import com.ajousw.spring.domain.navigation.exception.ApiNotSupportedException;
 import java.util.Map;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
