+++ src/main/java/com/ajousw/spring/domain/navigation/api/provider/impl/OsrmNavigationApi.java
@@ -1,7 +1,7 @@
 package com.ajousw.spring.domain.navigation.api.provider.impl;
 
-import com.ajousw.spring.domain.exception.BadApiResponseException;
 import com.ajousw.spring.domain.navigation.api.provider.NavigationApi;
+import com.ajousw.spring.domain.navigation.exception.BadApiResponseException;
 import jakarta.annotation.PostConstruct;
 import java.util.ArrayList;
 import java.util.List;
